+++ src/bzflag/defaultBZDB.cxx
@@ -147,7 +147,7 @@ DefaultDBItem	defaultDBItems[] = {
 
 void loadBZDBDefaults ( void )
 {
-	for (int i = 0; i < countof(defaultDBItems); ++i) 
+	for (int i = 0; i < (int)countof(defaultDBItems); ++i) 
 	{
 		assert(defaultDBItems[i].name != NULL);
